tests/wpt/tests/clipboard-apis/text-write-read/async-writeText-readText.https.html

<!doctype html>
<meta charset="utf-8">
<title>Async Clipboard writeText -> readText tests</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="../resources/user-activation.js"></script>
<script>
async function readWriteTest(textInput) {
  promise_test(async t => {
    await tryGrantReadPermission();
    await tryGrantWritePermission();

    await waitForUserActivation();
    await navigator.clipboard.writeText(textInput);
    await waitForUserActivation();
    const textOutput = await navigator.clipboard.readText();

    assert_equals(textOutput, textInput);
  }, 'Verify write and read clipboard given text: ' + textInput);
}

readWriteTest('Clipboard write text -> read text test');
readWriteTest('non-Latin1 text encoding test データ');
</script>
